CAY-821 EJBQL resolution: FrontBase limitations skipping failing unit tests
git-svn-id: https://svn.apache.org/repos/asf/cayenne/main/trunk@552319 13f79535-47bb-0310-9956-ffa450edef68

+++ framework/cayenne-jdk1.4-unpublished/src/test/java/org/apache/cayenne/access/DataContextEJBQLIsNullTest.java
@@ -28,6 +28,12 @@ import org.apache.cayenne.unit.CayenneCase;
 public class DataContextEJBQLIsNullTest extends CayenneCase {
 
     public void testCompareToNull() throws Exception {
+        // the query below can blow up on FrontBase. See CAY-819 for details.
+
+        if (!getAccessStackAdapter().supportsEqualNullSyntax()) {
+            return;
+        }
+
         deleteTestData();
         createTestData("prepare");
 
@@ -41,8 +47,6 @@ public class DataContextEJBQLIsNullTest extends CayenneCase {
         // most do not; per JPA spec the result is undefined.. so we can't make any
         // assertions about the result. Just making sure the query doesn't blow up
         createDataContext().performQuery(query1);
-        
-        // still this query can blow up on FrontBase. See CAY-819 for details.
     }
 
     public void testIsNull() throws Exception {
